pyproject.toml
[tool.poetry]
name = "azure-blob-handlers"
version = "0.1.0.dev-1"
description = "Azure Blob Handlers"
authors = ["osrokas <rokasptr@gmail.com>"]
repository = "https://github.com/osrokas/azure-handlers.git"
readme = ["README.md", "LICENSE"]
keywords =  ["azure", "handlers", "blob"]
packages = [
    { include = "blobhand", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.9"
pytest = "^8.3.3"
azure-storage-blob = "^12.24.0"
pandas = "^2.2.3"
geopandas = "^1.0.1"
geojson = "^3.1.0"
pyarrow = "^18.1.0"
fastparquet = "^2024.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"